+++ tests/machine/x/cs/fun_expr_in_let.cs
@@ -2,7 +2,7 @@ using System;
 
 class Program {
     static void Main() {
-        Func<int, int> square = new Func<long, long>((long x) => {
+        Func<int, int> square = new Func<int, int>((int x) => {
     return (x * x);
 });
         Console.WriteLine(square(6));
