+++ tests/machine/x/cs/let_and_print.cs
@@ -3,7 +3,7 @@ using System;
 class Program {
     static void Main() {
         int a = 10;
-        long b = 20;
+        int b = 20;
         Console.WriteLine((a + b));
     }
 }
