+++ tests/machine/x/cs/load_yaml.cs
@@ -6,7 +6,7 @@ using System.Text.Json;
 
 public record struct Person {
     public string name;
-    public long age;
+    public int age;
     public string email;
 }
