+++ tests/machine/x/cs/nested_function.cs
@@ -1,8 +1,8 @@
 using System;
 
 class Program {
-    static long outer(long x) {
-        long inner(long y) {
+    static int outer(int x) {
+        int inner(int y) {
             return (x + y);
         }
         return inner(5);
